Reorganise AppletFifo
