+++ Ryujinx.HLE/HOS/Services/Am/AppletFifo.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 
-namespace Ryujinx.HLE.HOS.Applets
+namespace Ryujinx.HLE.HOS.Services.Am
 {
     internal class AppletFifo<T> : IEnumerable<T>
     {
